moved app within its own directory
