+++ app.js
@@ -6,7 +6,7 @@ const morgan = require('morgan');
 const app = express();
 
 // view engine setup
-app.set('views', path.join(__dirname, 'views'));
+app.set('views', path.join(__dirname, 'app/views'));
 app.set('view engine', 'ejs');
 
 // middleware
@@ -14,17 +14,17 @@ app.use(morgan('dev'));
 app.use(express.urlencoded({ extended: true }));
 app.use(express.json());
 app.use(cookieParser());
-app.use(express.static(path.join(__dirname, 'public')));
+app.use(express.static(path.join(__dirname, 'app/public')));
 
 // app routes
-app.use('/', require('./routes/index'));
-app.use('/assess', require('./routes/assess'));
-app.use('/markets', require('./routes/markets'));
-app.use('/results', require('./routes/results'));
-app.use('/data', require('./routes/data'));
-app.use('/marketdata', require('./routes/marketdata'));
-app.use('/admin-login', require('./routes/admin-login'));
-app.use('/submit-assess', require('./routes/submit-assess'));
+app.use('/', require('./app/routes/index'));
+app.use('/assess', require('./app/routes/assess'));
+app.use('/markets', require('./app/routes/markets'));
+app.use('/results', require('./app/routes/results'));
+app.use('/data', require('./app/routes/data'));
+app.use('/marketdata', require('./app/routes/marketdata'));
+app.use('/admin-login', require('./app/routes/admin-login'));
+app.use('/submit-assess', require('./app/routes/submit-assess'));
 
 const port = 3000;
 app.listen(port, () => {
