+++ app/config/index.js
@@ -1,4 +1,4 @@
-config = {
+module.exports = {
     apiKey: "AIzaSyB33kxrp6W3dihuRRFhgk2UFI9VSHCQvtI",
     authDomain: "live-weller.firebaseapp.com",
     databaseURL: "https://live-weller.firebaseio.com",
@@ -6,5 +6,3 @@ config = {
     storageBucket: "live-weller.appspot.com",
     messagingSenderId: "147602049424"
 };
-
-exports.config = config;
