+++ app/public/js/bootstrap-dropdown.js
