+++ app/public/js/navbar-highlighter.js
