+++ app/public/js/navbar.js
