+++ package.json
@@ -14,5 +14,10 @@
 		"firebase": "^6.6.1",
 		"http-errors": "~1.6.2",
 		"morgan": "^1.9.0"
+	},
+	"devDependencies": {
+		"eslint": "^6.5.1",
+		"jest": "^24.9.0",
+		"nodemon": "^1.19.3"
 	}
 }
